Change thread_to_scap to do a shallow copy
A big cost in creating a new capture file from a live event stream is
dumping the current set of threads to the capture file. This is done
in sinsp_thread_manager::dump_threads_to_file() and involves iterating
over the set of threads and converting each to a scap_threadinfo
struct, multiple times in fact.

To help make that more efficient, we can do a shallow copy of the
strings from the sinsp_threadinfo struct given that the corresponding
scap_threadinfo is temporary.

This requires some additional management in sinsp_threadinfo to save a
pre-combined version of the process's arguments/environment/cgroups, but
also eliminates the need for {args,env,cgroups}_to_scap, as you don't
need to populate those fields in a scap_threadinfo any longer.