Add way to dump scap_threadinfo w/ alternate bufs
Add a new scap_write_proclist_entry_bufs which takes a scap_threadinfo
and a set of strings representing the various array-backed strings that
can be in a scap_threadinfo struct.

This is useful when dumping the set of (sinsp) threads, as you can dump
a thread without copying all of the strings into the scap_threadinfo
struct.