Reorganize ruleset mgmt in sinsp_evttype_filter
Change how rulesets are organized in sinsp_evttype_filter. Previously
the primary index into the set of filters was by event type. For filters
matching an event type, each would be considered to see if they were
relevant for a given ruleset, and if so, the filter would be run.

This ends up being inefficient for the use case of iterating over
rulesets, as you end up traversing the list of filters for a given event
type N times for N rulesets.

Instead, add a new object ruleset_filters which holds the set of filters
for a given ruleset. The lists of filters indexed by evttype/syscall
id move from the main object to the ruleset_filters object.

Also, get rid of the notion of "catchall" filters. If a filter really
should apply for all evttypes/syscalls, it just gets included in every
m_filter_by_{evttype,syscall} list. Catchall filters are highly
discouraged and don't exist for the default falco ruleset, so it's ok to
not have a special case for it. This ends up significantly simplifying
::run(), as you can iterate over a single list instead of two and don't
need to keep track of any order value for a given filter.